Apply demurrage (#153)
* Add chainApi
* display the true encointerBalance (principal - demurrage) everywhere in the app.
* Fix location type for dev chains with updated degree type: `I64F64`

Closes #133.

+++ lib/store/encointer/types/location.dart
@@ -9,8 +9,8 @@ part 'location.g.dart';
 class Location {
   Location(this.lon, this.lat);
 
-  final int lon;
-  final int lat;
+  final BigInt lon;
+  final BigInt lat;
 
   @override
   String toString() {
